+++ Backend/src/test/java/com/firomsa/inventory/v1/service/SaleServiceUnitTest.java
@@ -9,6 +9,7 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -59,7 +60,7 @@ public class SaleServiceUnitTest {
         // Arrange
         var productId = UUID.randomUUID();
         var request = SaleRequestDTO.builder().productId(productId).quantity(5)
-                .salePrice(10.0).build();
+                .salePrice(BigDecimal.valueOf(10.0)).build();
         var response = SaleResponseDTO.builder().productId(request.getProductId())
                 .quantity(request.getQuantity()).salePrice(request.getSalePrice())
                 .message("Sale recorded successfully").build();
@@ -90,7 +91,7 @@ public class SaleServiceUnitTest {
     void shouldThrowExceptionWhenProductNotFound() {
         // Arrange
         var request = SaleRequestDTO.builder().productId(UUID.randomUUID()).quantity(5)
-                .salePrice(10.0).build();
+                .salePrice(BigDecimal.valueOf(10.0)).build();
         when(productRepository.findById(request.getProductId())).thenReturn(Optional.empty());
 
         // Act & Assert
@@ -104,7 +105,7 @@ public class SaleServiceUnitTest {
     void shouldThrowExceptionWhenInsufficientStock() {
         // Arrange
         var request = SaleRequestDTO.builder().productId(UUID.randomUUID()).quantity(15)
-                .salePrice(10.0).build();
+                .salePrice(BigDecimal.valueOf(10.0)).build();
         var product = getProduct();
         var user = getUser();
         when(productRepository.findById(request.getProductId())).thenReturn(Optional.of(product));
@@ -123,14 +124,14 @@ public class SaleServiceUnitTest {
         var product = getProduct();
         var user = getUser();
         var firstSale = Sale.builder().id(UUID.randomUUID()).product(product).soldBy(user)
-                .quantity(2).salePrice(12.0).build();
+                .quantity(2).salePrice(BigDecimal.valueOf(12.0)).build();
         var secondSale = Sale.builder().id(UUID.randomUUID()).product(product).soldBy(user)
-                .quantity(4).salePrice(15.0).build();
+                .quantity(4).salePrice(BigDecimal.valueOf(15.0)).build();
 
         var firstResponse = SaleResponseDTO.builder().productId(product.getId()).quantity(2)
-                .salePrice(12.0).build();
+                .salePrice(BigDecimal.valueOf(12.0)).build();
         var secondResponse = SaleResponseDTO.builder().productId(product.getId()).quantity(4)
-                .salePrice(15.0).build();
+                .salePrice(BigDecimal.valueOf(15.0)).build();
 
         when(saleRepository.findAll()).thenReturn(List.of(firstSale, secondSale));
         when(saleMapper.toDTO(eq(firstSale))).thenReturn(firstResponse);
@@ -152,9 +153,9 @@ public class SaleServiceUnitTest {
         var product = getProduct();
         var user = getUser();
         var sale = Sale.builder().id(saleId).product(product).soldBy(user).quantity(3)
-                .salePrice(11.0).build();
+                .salePrice(BigDecimal.valueOf(11.0)).build();
         var response = SaleResponseDTO.builder().productId(product.getId()).quantity(3)
-                .salePrice(11.0).build();
+                .salePrice(BigDecimal.valueOf(11.0)).build();
 
         when(saleRepository.findById(saleId)).thenReturn(Optional.of(sale));
         when(saleMapper.toDTO(sale)).thenReturn(response);
@@ -164,7 +165,7 @@ public class SaleServiceUnitTest {
 
         // Assert
         assertThat(result.getQuantity()).isEqualTo(3);
-        assertThat(result.getSalePrice()).isEqualTo(11.0);
+        assertThat(result.getSalePrice()).isEqualTo(BigDecimal.valueOf(11.0));
     }
 
     @Test
@@ -187,9 +188,9 @@ public class SaleServiceUnitTest {
         var product = getProduct();
         var user = getUser();
         var sale = Sale.builder().id(UUID.randomUUID()).product(product).soldBy(user).quantity(6)
-                .salePrice(20.0).build();
+                .salePrice(BigDecimal.valueOf(20.0)).build();
         var response = SaleResponseDTO.builder().productId(product.getId()).quantity(6)
-                .salePrice(20.0).build();
+                .salePrice(BigDecimal.valueOf(20.0)).build();
 
         when(saleRepository.findBySoldByEmail(email)).thenReturn(List.of(sale));
         when(saleMapper.toDTO(sale)).thenReturn(response);
@@ -230,7 +231,7 @@ public class SaleServiceUnitTest {
     void shouldTriggerNotificationWhenLowStockAfterSale() {
         // Arrange
         var request = SaleRequestDTO.builder().productId(UUID.randomUUID()).quantity(8)
-                .salePrice(10.0).build();
+                .salePrice(BigDecimal.valueOf(10.0)).build();
         var response = SaleResponseDTO.builder().productId(request.getProductId())
                 .quantity(request.getQuantity()).salePrice(request.getSalePrice())
                 .message("Sale recorded successfully").build();
@@ -258,7 +259,7 @@ public class SaleServiceUnitTest {
     void shouldNotTriggerNotificationWhenStockStillSufficient() {
         // Arrange
         var request = SaleRequestDTO.builder().productId(UUID.randomUUID()).quantity(2)
-                .salePrice(10.0).build();
+                .salePrice(BigDecimal.valueOf(10.0)).build();
         var response = SaleResponseDTO.builder().productId(request.getProductId())
                 .quantity(request.getQuantity()).salePrice(request.getSalePrice())
                 .message("Sale recorded successfully").build();
@@ -286,7 +287,7 @@ public class SaleServiceUnitTest {
     void createSale_WhenEmailNotificationFails_ShouldStillCompleteSaleTransaction() {
         // Arrange
         var request = SaleRequestDTO.builder().productId(UUID.randomUUID()).quantity(3)
-                .salePrice(15.0).build();
+                .salePrice(BigDecimal.valueOf(15.0)).build();
         var response = SaleResponseDTO.builder().productId(request.getProductId())
                 .quantity(request.getQuantity()).salePrice(request.getSalePrice())
                 .message("Sale recorded successfully").build();
